+++ pkg/kubectl/cmd/create_service_test.go
@@ -23,12 +23,13 @@ import (
 
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/client/unversioned/fake"
+	cmdtesting "k8s.io/kubernetes/pkg/kubectl/cmd/testing"
 )
 
 func TestCreateService(t *testing.T) {
 	service := &api.Service{}
 	service.Name = "my-service"
-	f, tf, codec, negSer := NewAPIFactory()
+	f, tf, codec, negSer := cmdtesting.NewAPIFactory()
 	tf.Printer = &testPrinter{}
 	tf.Client = &fake.RESTClient{
 		NegotiatedSerializer: negSer,
